requirements.txt
requests==2.31.0
boto3==1.34.0
python-dotenv==1.0.0
pandas==2.1.0
pyarrow==14.0.0
redshift-connector==2.0.918